Restructure project and implement lore fetching
