--- api/alienworlds.js
@@ -1,5 +1,0 @@
-export async function fetchCanonLore() {
-  const res = await fetch('https://alienworlds-api.com/lore/canon');
-  const data = await res.json();
-  return data;
-}
