+++ components/votingSystem.js
